Add 2nd version of XLS > SRT
This script was inspired by a new script format for an AD company I will not name, which was part of a series of projects I had to work on.

Due to changes in the software writers were using, this new format threw off version 1 of this script. The column names MUST match, meaning, if there are two "event" columns, but one of them are just numbers, change that one to "Event Number". The script should then work.

Like with all of these scripts, feel free to tweak to your own needs.